* commit '468a514ab376ae81da3ba24fa1a2c50c9c2bd8ac':
  Lower dmg
  Slow fire rate
  Fix ...
  Refactor enemy system setMap()
  reset particles in pool
  Keep particles
  Add GRAVITY const
  Fix coin count
  [WIP]Enable reset button

+++ src/Systems/death_system.cpp
@@ -2,25 +2,24 @@
 
 void DeathSystem::handleDeaths(EntityManager &entityManager)
 {
-	vector<shared_ptr<Entity>> dead = entityManager.getEntities(entityManager.getComponentChecker(vector<int> { ComponentType::death }));
-	for (shared_ptr<Entity> e : dead) {
-		TransformComponent *transform = e->getComponent<TransformComponent>();
-		EnemyComponent *enemy = e->getComponent<EnemyComponent>();	
-		ProjectileComponent *projectile = e->getComponent<ProjectileComponent>();
+  vector<shared_ptr<Entity>> dead = entityManager.getEntities(entityManager.getComponentChecker(vector<int>{ ComponentType::death }));
+  for (shared_ptr<Entity> e : dead) {
 
-		if (enemy != nullptr) {
-			for (Entity resource : ResourceFactory::buildCluster(enemy->worth, transform->position, transform->size)) {;
-				entityManager.addEntity(resource);
-			}
-			entityManager.removeEntity(e);
-			WavesetSystem::getInstance().decrementEnemies(1, entityManager);
-			ParticleSystem::emitParticleCluster(entityManager, transform->position);
-			Mix_PlayChannel(-1, AudioLoader::getInstance().enemy_dead, 0);
-		}
+    TransformComponent *transform = e->getComponent<TransformComponent>();
+    EnemyComponent *enemy = e->getComponent<EnemyComponent>();
+    if (enemy != nullptr && transform != nullptr) {
+      for (Entity resource : ResourceFactory::buildCluster(enemy->worth, transform->position, transform->size)) {
+        entityManager.addEntity(resource);
+      }
+      entityManager.removeEntity(e);
+      WavesetSystem::getInstance().decrementEnemies(1, entityManager);
+      ParticleSystem::emitParticleCluster(entityManager, transform->position);
+      Mix_PlayChannel(-1, AudioLoader::getInstance().enemy_dead, 0);
+    }
 
-		if (projectile != nullptr) {
-			entityManager.removeEntity(e);
-		}
-
-	}
+    ProjectileComponent *projectile = e->getComponent<ProjectileComponent>();
+    if (projectile != nullptr) {
+      entityManager.removeEntity(e);
+    }
+  }
 }
